==> [Project] NBA Players Regression GSW 15-16/Getting the Data/2 Scrape RPM per Player.ipynb <==
# Run this code to get the dataset needed with the positions and the names.

### Functions

from bs4 import BeautifulSoup
import requests
from HTMLParser import HTMLParser
from datetime import datetime
import hashlib
import requests
import re

import numpy as np
import pandas as pd

class MLStripper(HTMLParser):
    def __init__(self):
        self.reset()
        self.strict = False
        self.convert_charrefs= True
        self.fed = []
    def handle_data(self, d):
        self.fed.append(d)
    def get_data(self):
        return ''.join(self.fed)

def strip_tags(html):
    s = MLStripper()
    s.feed(html)
    return s.get_data()


# In[58]:
from selenium import webdriver
import urllib2
import MySQLdb

# MAIN


### Read the NBA File you generated from the Appendix code below.

NBA_dataset = pd.read_csv("NBA_teamnamepos_data.csv", sep=",")
NBA_dataset.head()

## Go to the ESPN site and get the data you want, which is the RPM info. 

#initialize data frame
RPM_data = pd.DataFrame(columns=('names', 'GP','MPG', 'ORPM', 'DRPM', 'RPM', 'WINS'))

print str(datetime.now())
for j in range(0, 12):
    
    url = "http://www.espn.com/nba/statistics/rpm/_/year/2016/page/" + str(j) + "/sort/RPM"

    print url
    
    r1 = urllib2.urlopen(url)
    soup = BeautifulSoup(r1, 'html.parser')


    #get the even columns
    evens = soup.findAll('tr', {'class':'evenrow'})

    for i in range(0, len(evens)):
        name = evens[i].a.text
        name

        team = evens[i].find("td").find_next_sibling("td").find_next_sibling("td")
        team.text

        games_played = team.find_next_sibling("td")
        games_played.text

        mins_per_game = games_played.find_next_sibling("td")
        mins_per_game.text

        ORPM = mins_per_game.find_next_sibling("td")
        ORPM.text

        DRPM = ORPM.find_next_sibling("td")
        DRPM.text

        RPM = DRPM.find_next_sibling("td")
        RPM.text

        wins = RPM.find_next_sibling("td")
        wins.text

        record = [name, games_played.text, mins_per_game.text, ORPM.text, DRPM.text, RPM.text, wins.text]

        temp_df = pd.DataFrame([record],columns=('names', 'GP','MPG', 'ORPM', 'DRPM', 'RPM', 'WINS'))
        frames = [RPM_data, temp_df]
        RPM_data = pd.concat(frames, ignore_index = True)


    #get the odd columns
    odds = soup.findAll('tr', {'class':'oddrow'})

    for i in range(0, len(odds)):
        name = odds[i].a.text
        name

        team = odds[i].find("td").find_next_sibling("td").find_next_sibling("td")
        team.text

        games_played = team.find_next_sibling("td")
        games_played.text

        mins_per_game = games_played.find_next_sibling("td")
        mins_per_game.text

        ORPM = mins_per_game.find_next_sibling("td")
        ORPM.text

        DRPM = ORPM.find_next_sibling("td")
        DRPM.text

        RPM = DRPM.find_next_sibling("td")
        RPM.text

        wins = RPM.find_next_sibling("td")
        wins.text

        record = [name, games_played.text, mins_per_game.text, ORPM.text, DRPM.text, RPM.text, wins.text]

        temp_df = pd.DataFrame([record],columns=('names', 'GP','MPG', 'ORPM', 'DRPM', 'RPM', 'WINS'))
        frames = [RPM_data, temp_df]
        RPM_data = pd.concat(frames, ignore_index = True)

        #concatenate the odds and evens
    print len(RPM_data)
print str(datetime.now())

print len(RPM_data)
RPM_dataset = RPM_data.drop_duplicates(keep='first')
print len(RPM_dataset)

print len(NBA_dataset)

merged_data = pd.merge(NBA_dataset, RPM_dataset, how = 'left', on='names')
len(merged_data)

# save to a csv file 
merged_data.to_csv("NBA Dataset with RPM.csv", sep=',',index=False)

merged_data.head()

### Note that there are players with no recorded RPM; this is because the player names and the RPM info came from different sources.
