==> [Project] NBA Players Regression GSW 15-16/Getting the Data/4 Check Names and Positions.ipynb <==
# This code checks if the names and the positions match.

### Functions

from bs4 import BeautifulSoup
import requests
from HTMLParser import HTMLParser
from datetime import datetime
import hashlib
import requests
import re

import numpy as np
import pandas as pd

class MLStripper(HTMLParser):
    def __init__(self):
        self.reset()
        self.strict = False
        self.convert_charrefs= True
        self.fed = []
    def handle_data(self, d):
        self.fed.append(d)
    def get_data(self):
        return ''.join(self.fed)

def strip_tags(html):
    s = MLStripper()
    s.feed(html)
    return s.get_data()


# In[58]:
from selenium import webdriver
import urllib2
import MySQLdb

# MAIN


### Read the files

NBA_dataset = pd.read_csv("NBA Dataset with RPM.csv", sep=",")
NBA_dataset.head()

positions_DS = pd.read_csv("Pbp Data Positions.csv", sep=",")
positions_DS.head()

names_DS = pd.read_csv("Pbp Data Names.csv", sep=",")
names_DS.head()

# make a dict for GSW for faster searching
from datetime import datetime 
dict_GSW = {}
errors_list  = []

print str(datetime.now())

for i,row in names_DS.iterrows():
    for k in range(1,11):
        if row[k] not in dict_GSW.keys():
            
            # get the official position (NBA)
            pos = NBA_dataset[NBA_dataset['names'] == row[k]].iloc[0]['pos']
            
            dict_GSW[row[k]] = pos
        else:
            #do the checking
            pos_at_posDS = positions_DS.iloc[i][k]


            if dict_GSW[row[k]] == pos_at_posDS:
                pass
            else:
                #maybe they are traded
                pos2 = NBA_dataset[NBA_dataset['names'] == row[k]].iloc[1]['pos']
                dict_GSW[row[k]] = pos2
                if dict_GSW[row[k]] == pos_at_posDS:
                    pass
                else:
                    errors_list.append(row[k])

        
print str(datetime.now())

print "Finished checking."
if len(errors_list) == 0:
    print "No Errors."
else:
    print "There are ", len(errors_list), " errors."
