==> [Project] NBA Players Regression GSW 15-16/Getting the Data/5 Generate a More Useable Positions Dataset.ipynb <==
## Convert the Ultra Pbp Positions to a more useable format for regression.

### Functions

from bs4 import BeautifulSoup
import requests
from HTMLParser import HTMLParser
from datetime import datetime
import hashlib
import requests
import re

import numpy as np
import pandas as pd

class MLStripper(HTMLParser):
    def __init__(self):
        self.reset()
        self.strict = False
        self.convert_charrefs= True
        self.fed = []
    def handle_data(self, d):
        self.fed.append(d)
    def get_data(self):
        return ''.join(self.fed)

def strip_tags(html):
    s = MLStripper()
    s.feed(html)
    return s.get_data()


# In[58]:
from selenium import webdriver
import urllib2
import MySQLdb

# MAIN


### Read the files

positions_DS = pd.read_csv("Pbp Data Positions.csv", sep=",")
positions_DS.head()

positions = []
dataset = pd.DataFrame(columns=('id','SFnum_G', 'PFnum_G', 'Cnum_G', 'SGnum_G', 'PGnum_G', 'SFnum_O', 'PFnum_O', 'Cnum_O', 'SGnum_O', 'PGnum_O', 'secs', 'score_GSW', 'score_OPP', 'sO-sG','game'))
print str(datetime.now())
for i,row in positions_DS.iterrows():
    dict_pos_GSW = {'SF':0, 'PF':0, 'C':0, 'SG':0, 'PG':0}
    dict_pos_OPP = {'SF':0, 'PF':0, 'C':0, 'SG':0, 'PG':0}

    
    id_val = []
    end_values = []
    #iterate per row
    for j in range(0,16):
        if j == 0:
            id_val.append(int(row[j]))
        elif j >= 1 and j <= 5:
            dict_pos_GSW[row[j]] += 1
        elif j >= 6 and j <= 10:
            dict_pos_OPP[row[j]] += 1
        elif j != 15:
            end_values.append(int(row[j]))
        else:
            end_values.append(row[j])
    #record the
    record = id_val + [dict_pos_GSW['SF'], dict_pos_GSW['PF'], dict_pos_GSW['C'], dict_pos_GSW['PG'], dict_pos_GSW['SG']]
    record = record + [dict_pos_OPP['SF'], dict_pos_OPP['PF'],dict_pos_OPP['C'],dict_pos_OPP['PG'],dict_pos_OPP['SG']] + end_values
    #print record

    temp = pd.DataFrame([record], columns=('id','SFnum_G', 'PFnum_G', 'Cnum_G', 'SGnum_G', 'PGnum_G', 'SFnum_O', 'PFnum_O', 'Cnum_O', 'SGnum_O', 'PGnum_O', 'secs', 'score_GSW', 'score_OPP', 'sO-sG','game'))
    frames = [dataset, temp]
    dataset = pd.concat(frames, ignore_index = True)

print str(datetime.now())


dataset.to_csv("Pbp Positions for Regression.csv", sep=',',index=False)

dataset.head(51)

## After this, we use the "Pbp Positions for Regression.csv" and the "NBA Dataset with RPM.csv" to help us do regression and figure out which players from the opponent's side has the best chances of defeating the GSW.
